+++ .gitignore
@@ -1,3 +1,4 @@
 *.pyc
 **/*.pyc
 .DS_store
+database
