+++ blockchain/Dockerfile
@@ -1,5 +1,8 @@
 FROM python:3.6.1-alpine
 WORKDIR /blockchain
-ADD . /blockchain
+COPY  requirements.txt ./
 RUN pip install -r requirements.txt
+
+ADD . /blockchain
+
 CMD ["python","run_app.py"]
