+++ node_server/__init__.py
@@ -1,1 +1,2 @@
+from .node_server import db
 from .node_server import create_app
