+++ node_server/entrypoint/init.sh
@@ -1,7 +1,4 @@
 #!bin/sh
 
-#initalice database
-python3 ./database/init_database.py
-
 #Run flask app
-FLASK_APP=node_server.py flask run --port=8000 --host=0.0.0.0
+./wait_for_mysql.sh && FLASK_APP=node_server.py flask run --port=8000 --host=0.0.0.0
