+++ node_server/requirements.txt
@@ -1,3 +1,5 @@
 Flask~=1.1
 requests~=2.22
 jsons==1.2.0
+Flask-SQLAlchemy==2.4.4
+PyMySQL==0.10.1
