+++ node_server/wait_for_mysql.sh
@@ -1,0 +1,10 @@
+#!/bin/sh
+
+while ! mysqladmin ping -h"mysql_database" -P"3306" --silent; do
+    echo "Waiting for MySQL to be up..."
+    sleep 1
+done
+
+sleep 3
+echo "Starting node_server..."
+exit 0
